+++ app/frontend/src/components/Keyboard/layouts/index.ts
@@ -1,0 +1,8 @@
+import type { Keyboard } from '../types'
+import { Halo75 } from './halo75'
+import { Halo65 } from './halo65'
+
+export const layouts: Record<string, Keyboard> = {
+  Halo75,
+  Halo65
+}